[PodLevelResources] Update Downward API defaulting for resource limits
Currently, when container-level resource limits were not specified and
the Downward API was used to set environment variables referencing them,
the node's allocatable resources were used as the fallback.
With the introduction of the Pod Level Resources feature, this behavior
is updated: if container-level resource limits are not specified,
the Downward API now uses the pod-level resource limits instead.
If neither container-level nor pod-level resource limits are specified,
the behavior remains unchanged. It falls back to the node's allocatable
resources.

Signed-off-by: Tsubasa Nagasawa <toversus2357@gmail.com>

Kubernetes-commit: a82187cf11aa41e0f907761bdd4b815984f6de67